backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langchain-pinecone
pymupdf
pydantic
pinecone-client